% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R, R/color_palette_euf.R
\name{coloring}
\alias{coloring}
\alias{scale_color_icae}
\alias{scale_fill_icae}
\alias{scale_color_euf}
\alias{scale_fill_euf}
\title{Color scale in the ICAE color scheme.}
\usage{
scale_color_icae(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_icae(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_color_euf(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_euf(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The type of palette to be returned. Currently, the follwoing
palettes are supported: \code{main}, \code{cool}, \code{hot}, and
\code{mixed}.}

\item{discrete}{If TRUE returnes a discrete scheme.}

\item{reverse}{If TRUE reverses the resulting color scheme.}
}
\description{
Creates a color scale in accordance with the official ICAE color scheme.

Creates a color scale in accordance with the official EUF color scheme.
}
\details{
These functions can be used to transform the color scheme of a \code{ggplot}
 so that it is in accordance with the official ICAE color scheme.
 The function provides functionality for continuous and discrete color
 schemes but has still problems in handling plots with too many different
 colors. For the \code{colour} aesthetic use \code{scale_color_icae}, for
 the \code{fill} aesthetic use \code{scale_fill_icae}.

 The palettes currently available are the following:

 \code{main}: "sand", "purple", "dark red"

 \code{cool}: "purple", "dark green", "dark blue"

 \code{hot}: "sand", "dark red"

 \code{mixed}: "orange", "dark blue", "purple", "sand", "dark red"

 \code{grey}: "light grey", "dark grey"

These functions can be used to transform the color scheme of a \code{ggplot}
 so that it is in accordance with the official EUF color scheme.
 The function provides functionality for continuous and discrete color
 schemes but has still problems in handling plots with too many different
 colors. For the \code{colour} aesthetic use \code{scale_color_euf}, for
 the \code{fill} aesthetic use \code{scale_fill_euf}.

 The palettes currently available are the following:

 \code{main}: "grey", "blue"

 \code{cool}: "green", "lightblue", "blue"

 \code{hot}: "grey", "red"

 \code{mixed}: "blue", "red", "green", "lightblue", "grey"

 \code{grey}: "light grey", "dark grey"
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
 geom_point(size = 4) +
 scale_color_icae("hot")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
 geom_point(size = 4, alpha = .6) +
 scale_color_icae(discrete = FALSE, palette = "cool")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
 geom_point(size = 4) +
 scale_color_euf("hot")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
 geom_point(size = 4, alpha = .6) +
 scale_color_euf(discrete = FALSE, palette = "cool")

}
\seealso{
The package was loosely built upon the explanations in
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}.

The package was loosely built upon the explanations in
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}.

Other color scheme functions: 
\code{\link{get_euf_colors}()},
\code{\link{get_icae_colors}()}

Other color scheme functions: 
\code{\link{get_euf_colors}()},
\code{\link{get_icae_colors}()}
}
\concept{color scheme functions}
