% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette_euf.R
\name{euf_public_cols}
\alias{euf_public_cols}
\title{Get hex codes for colors}
\usage{
euf_public_cols(...)
}
\description{
Contains the hex codes for all the colors that are part of the color scheme.
}
\seealso{
\code{\link{euf_public_pal}} for the function that assembles the
 colors into consistent palettes and \code{\link{scale_color_euf}} for the
 application to \code{ggplot2} objects. See \code{\link{icae_public_pal}}
 and \code{\link{scale_color_icae}} for the corresponding functions for the
 ICAE case.
}
