% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette_euf.R
\name{get_euf_colors}
\alias{get_euf_colors}
\title{Get EUF colors}
\usage{
get_euf_colors(
  n = 1,
  col_name = NULL,
  palette_used = "main",
  reverse_pal = FALSE,
  ...
)
}
\arguments{
\item{n}{The number of different colors requested.}

\item{col_name}{The name of a color as defined in
\code{\link{euf_public_cols}}. If 'all' returns all EUF colors.}

\item{palette_used}{The type of palette to be returned. Currently, the
following palettes are supported: \code{main}, \code{cool}, \code{hot},
and  \code{mixed}.}

\item{reverse_pal}{If TRUE reverses the resulting color scheme.}
}
\value{
A vector of hex codes with colors from \code{palette}
 (if \code{col_name} is \code{FALSE}), a named vector of hex codes from
 the colors specified in \code{col_name} otherwise.
}
\description{
Returns a vector of colors of a chose EUF-color palette.
}
\details{
This function takes one of the palettes defined in
 \code{\link{euf_public_pal}} and returns a vector with the requested
 number of colors. To return all base colors of the EUF color scheme
 call the function with \code{col_name="all"}.
}
\examples{
get_euf_colors(3)

get_euf_colors(2, palette_used = "mixed", reverse_pal = TRUE)

get_euf_colors("lightblue")

get_euf_colors(c("blue", "red"))

get_euf_colors(col_name="all")

}
\seealso{
\code{\link{euf_public_pal}} for the function that assembles the
 colors into consistent palettes and \code{\link{scale_color_euf}} for the
 application to \code{ggplot2} objects. See \code{\link{icae_public_pal}}
 and \code{\link{scale_color_euf}} for the ICAE pendants.

Other color scheme functions: 
\code{\link{coloring}},
\code{\link{get_icae_colors}()}
}
\concept{color scheme functions}
