% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{get_icae_colors}
\alias{get_icae_colors}
\title{Get ICAE colors}
\usage{
get_icae_colors(n, palette_used = "main", reverse_pal = FALSE, ...)
}
\arguments{
\item{n}{The number of different colors requested.}

\item{palette_used}{The type of palette to be returned. Currently, the
follwoing palettes are supported: \code{main}, \code{cool}, \code{hot},
\code{mixed}, and  \code{grey}.}

\item{reverse_pal}{If TRUE reverses the resulting color scheme.}
}
\value{
A vector of hex codes with colors from \code{palette}.
}
\description{
Returns a vector of colors of a chose ICAE-color palette.
}
\details{
This function takes one of the palettes defined in
 \code{\link{icae_public_pal}} and returnes a vector with the requested
 number of colors.
}
\examples{
get_icae_colors(3)

get_icae_colors(2, palette_used = "mixed", reverse_pal = TRUE)

}
\seealso{
\code{\link{icae_public_pal}} for the function that assembles the
 colors into consistent palettes and \code{\link{scale_color_icae}} for the
 application to \code{ggplot2} objects.

Other color scheme functions: \code{\link{coloring}}
}
\concept{color scheme functions}
