% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{icae_public_cols}
\alias{icae_public_cols}
\title{Get hex codes for colors}
\usage{
icae_public_cols(...)
}
\description{
Contains the hex codes for all the colors that are part of the color scheme.
}
\seealso{
\code{\link{icae_public_pal}} for the function that assembles the
 colors into consistent palettes and \code{\link{scale_color_icae}} for the
 application to \code{ggplot2} objects.
}
