% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icaeDesign.R
\docType{package}
\name{icaeDesign}
\alias{icaeDesign}
\title{icaeDesign: A package to produce figures in the ICAE corporate design style.}
\description{
The package provides function to create figures in accordance withe the
 corporate design of the Institute for Comprehensive Analysis of the Economy
 (ICAE) at the Johannes Kepler University in Linz.
}
\section{Functions to update ggplot2 objects}{

 The functions \code{\link{scale_color_icae}} and
 \code{\link{scale_color_euf}}, as well as
 \code{\link{scale_fill_icae}} and \code{\link{scale_fill_euf}} return color
 schemes for the color and fill aesthetic in ggplots using the respective
 color schemes.
}

\seealso{
The package was loosely built upon the explanations in
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}.
}
