% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_icae}
\alias{theme_icae}
\title{A ggplot theme for the icaeDesign package}
\usage{
theme_icae(base_size = 11, base_family = "",
  base_line_size = base_size/170, base_rect_size = base_size/170)
}
\description{
A theme summarizing the typical adjustments made to ggplots.
}
\details{
This theme is not specifically tuned to the ICAE color scheme, but
 summarizes typical modifications that improve readability of a
 ggplot. It is in principle suitable for all kinds of plots and has
 a decent touch of blue/grey colors for lines and fonts.
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
 geom_point(size = 4) +
 scale_color_icae("hot") +
 theme_icae()
}
\concept{ggplot themes}
